Add prompt suggestions and chat grouping, fix bugs

+++ Hackchat/Core/MainView.swift
@@ -12,7 +12,7 @@ struct MainView: View {
     
     @ObservedObject var coreDataManager = CoreDataManager.shared
     
-    @FetchRequest(entity: Chat.entity(), sortDescriptors: [NSSortDescriptor(keyPath: \Chat.timestamp, ascending: false)]) var chats: FetchedResults<Chat>
+    @FetchRequest(entity: Chat.entity(), sortDescriptors: [NSSortDescriptor(keyPath: \Chat.lastEdited, ascending: false)]) var chats: FetchedResults<Chat>
     
     @State private var hasPaused = false
     @State private var isSearching = false
@@ -37,6 +37,12 @@ struct MainView: View {
             }
     }
     
+    private func groupedChats(_ chats: [Chat]) -> [ChatDateGroup: [Chat]] {
+        Dictionary(grouping: chats) { chat in
+            ChatDateGroup.group(for: chat.lastEdited ?? (chat.timestamp ?? Date()))
+        }
+    }
+    
     var body: some View {
         NavigationStack(path: $path) {
             VStack(spacing: 0) {
@@ -45,7 +51,7 @@ struct MainView: View {
                         isSearching.toggle()
                         isSearchFocused = isSearching
                     }
-                    .opacity(filteredChats.isEmpty ? 0 : 1)
+                    .opacity((filteredChats.isEmpty && !isSearching) ? 0 : 1)
                     .animation(animation, value: filteredChats.isEmpty)
                     NavigationBarButton("square.and.pencil", alignment: .trailing) {
                         let newChat = Chat(context: viewContext)
@@ -90,18 +96,27 @@ struct MainView: View {
                 } else {
                     ScrollView {
                         VStack {
-                            ForEach(filteredChats) { chat in
-                                NavigationLink(value: chat) {
-                                    ChatRowView(chat: chat)
-                                        .transition(.opacity)
-                                }
-                                .contentShape(.contextMenuPreview, .rect(cornerRadius: 15))
-                                .contextMenu {
-                                    Button(role: .destructive) {
-                                        selectedChat = chat
-                                        showChatDeleteConfirmation = true
-                                    } label: {
-                                        Label("Delete", systemImage: "trash")
+                            ForEach(ChatDateGroup.allCases, id: \.self) { group in
+                                if let chats = groupedChats(filteredChats)[group] {
+                                    VStack(alignment: .leading, spacing: 8) {
+                                        Text(group.rawValue)
+                                            .font(.system(size: 16))
+                                            .foregroundStyle(.gray)
+                                        ForEach(chats, id: \.id) { chat in
+                                            NavigationLink(value: chat) {
+                                                ChatRowView(chat: chat)
+                                                    .transition(.opacity)
+                                            }
+                                            .contentShape(.contextMenuPreview, .rect(cornerRadius: 15))
+                                            .contextMenu {
+                                                Button(role: .destructive) {
+                                                    selectedChat = chat
+                                                    showChatDeleteConfirmation = true
+                                                } label: {
+                                                    Label("Delete", systemImage: "trash")
+                                                }
+                                            }
+                                        }
                                     }
                                 }
                             }
@@ -130,7 +145,8 @@ struct MainView: View {
             })
             .onAppear {
                 for chat in chats {
-                    if chat.messages.array().isEmpty {
+                    // Check lastEdited because we don't want to delete the chat if the user edited the custom instructions or title
+                    if chat.messages.array().isEmpty && chat.lastEdited == nil {
                         // This deletes the chat when the user creates it and then doesn't use it
                         viewContext.delete(chat)
                         coreDataManager.save()
@@ -152,11 +168,15 @@ struct MainView: View {
 }
 
 struct ChatRowView: View {
-    var chat: Chat
+    @ObservedObject var chatViewModel: ChatViewModel
+    
+    init(chat: Chat) {
+        self.chatViewModel = ChatViewModel(chat)
+    }
     
     var body: some View {
         HStack {
-            Text(chat.name ?? "Untitled")
+            Text(chatViewModel.chat.name ?? "Untitled")
                 .foregroundStyle(Color.primary)
                 .fontWeight(.semibold)
             Spacer()
